+++ app/components/UI/Navbar/index.js
@@ -50,7 +50,6 @@ import { CommonSelectorsIDs } from '../../../../e2e/selectors/Common.selectors';
 import { WalletViewSelectorsIDs } from '../../../../e2e/selectors/WalletView.selectors';
 import { NetworksViewSelectorsIDs } from '../../../../e2e/selectors/Settings/NetworksView.selectors';
 import { SendLinkViewSelectorsIDs } from '../../../../e2e/selectors/SendLinkView.selectors';
-import { getBlockaidTransactionMetricsParams } from '../../../util/blockaid';
 
 const trackEvent = (event, params = {}) => {
   MetaMetrics.getInstance().trackEvent(event, params);
@@ -513,7 +512,6 @@ export function getSendFlowTitle(
   route,
   themeColors,
   resetTransaction,
-  transaction,
 ) {
   const innerStyles = StyleSheet.create({
     headerButtonText: {
@@ -529,12 +527,9 @@ export function getSendFlowTitle(
   });
   const rightAction = () => {
     const providerType = route?.params?.providerType ?? '';
-    const additionalTransactionMetricsParams =
-      getBlockaidTransactionMetricsParams(transaction);
     trackEvent(MetaMetricsEvents.SEND_FLOW_CANCEL, {
       view: title.split('.')[1],
       network: providerType,
-      ...additionalTransactionMetricsParams,
     });
     resetTransaction();
     navigation.dangerouslyGetParent()?.pop();
