+++ app/components/Views/confirmations/components/ApproveTransactionReview/index.js
@@ -51,7 +51,7 @@ import TransactionReviewDetailsCard from '../TransactionReview/TransactionReview
 import AppConstants from '../../../../../core/AppConstants';
 import { UINT256_HEX_MAX_VALUE } from '../../../../../constants/transaction';
 import { WALLET_CONNECT_ORIGIN } from '../../../../../util/walletconnect';
-import { getBlockaidTransactionMetricsParams } from '../../../../../util/blockaid';
+import { getBlockaidMetricsParams } from '../../../../../util/blockaid';
 import { withNavigation } from '@react-navigation/compat';
 import {
   isTestNet,
@@ -510,6 +510,21 @@ class ApproveTransactionReview extends PureComponent {
     clearInterval(intervalIdForEstimatedL1Fee);
   };
 
+  withBlockaidMetricsParams = () => {
+    let blockaidParams = {};
+
+    const { transaction } = this.props;
+    if (
+      transaction.id === transaction.currentTransactionSecurityAlertResponse?.id
+    ) {
+      blockaidParams = getBlockaidMetricsParams(
+        transaction.currentTransactionSecurityAlertResponse?.response,
+      );
+    }
+
+    return blockaidParams;
+  };
+
   getAnalyticsParams = () => {
     try {
       const { chainId, transaction, onSetAnalyticsParams } = this.props;
@@ -682,10 +697,9 @@ class ApproveTransactionReview extends PureComponent {
   };
 
   onContactUsClicked = () => {
-    const { transaction } = this.props;
     const analyticsParams = {
       ...this.getAnalyticsParams(),
-      ...getBlockaidTransactionMetricsParams(transaction),
+      ...this.withBlockaidMetricsParams(),
       external_link_clicked: 'security_alert_support_link',
     };
     this.props.metrics.trackEvent(
@@ -1163,13 +1177,13 @@ class ApproveTransactionReview extends PureComponent {
   };
 
   onCancelPress = () => {
-    const { onCancel, transaction } = this.props;
+    const { onCancel } = this.props;
     onCancel && onCancel();
     this.props.metrics.trackEvent(
       MetaMetricsEvents.APPROVAL_PERMISSION_UPDATED,
       {
         ...this.getAnalyticsParams(),
-        ...getBlockaidTransactionMetricsParams(transaction),
+        ...this.withBlockaidMetricsParams(),
       },
     );
   };
@@ -1186,7 +1200,7 @@ class ApproveTransactionReview extends PureComponent {
         MetaMetricsEvents.APPROVAL_PERMISSION_UPDATED,
         {
           ...this.getAnalyticsParams(),
-          ...getBlockaidTransactionMetricsParams(this.props.transaction),
+          ...this.withBlockaidMetricsParams(),
         },
       );
       return this.setState({ isReadyToApprove: true });
